PrometheusNaming changes were incompatible - so deprecating instead
