+++ prometheus-metrics-core/src/main/java/io/prometheus/metrics/core/metrics/Counter.java
@@ -230,7 +230,7 @@ public class Counter extends StatefulMetric<CounterDataPoint, Counter.DataPoint>
      * Prometheus.
      *
      * <p>Throws an {@link IllegalArgumentException} if {@link
-     * io.prometheus.metrics.model.snapshots.PrometheusNaming#isValidMetricName(String)
+     * io.prometheus.metrics.model.snapshots.PrometheusNames#isValidMetricName(String)
      * MetricMetadata.isValidMetricName(name)} is {@code false}.
      */
     @Override
