+++ prometheus-metrics-core/src/main/java/io/prometheus/metrics/core/metrics/Info.java
@@ -135,7 +135,7 @@ public class Info extends MetricWithFixedMetadata {
      * "runtime_info"} in Prometheus.
      *
      * <p>Throws an {@link IllegalArgumentException} if {@link
-     * io.prometheus.metrics.model.snapshots.PrometheusNaming#isValidMetricName(String)
+     * io.prometheus.metrics.model.snapshots.PrometheusNames#isValidMetricName(String)
      * MetricMetadata.isValidMetricName(name)} is {@code false}.
      */
     @Override
